requirements.txt
absl-py==1.4.0
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.1
django-filter==23.2
django-import-export==3.2.0
django-storages==1.13.2
et-xmlfile==1.1.0
executing==1.2.0
flatbuffers==23.5.9
fonttools==4.39.4
gast==0.4.0
google-api-core==2.11.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
jax==0.4.10
jedi==0.18.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
keras-facenet==0.3.2
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mtcnn==0.1.1
nest-asyncio==1.5.6
numpy==1.23.5
oauthlib==3.2.2
odfpy==1.4.1
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==4.23.0
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.4.0
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
utils==1.0.1
wcwidth==0.2.6
Werkzeug==2.3.4
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
